Moving each tests to its  folder
